Run `cargo +nightly clippy --fix`

+++ tests/by-util/test_mv.rs
@@ -621,7 +621,7 @@ fn test_mv_update_option() {
     at.touch(file_b);
     let ts = time::OffsetDateTime::now_local().unwrap();
     let now = FileTime::from_unix_time(ts.unix_timestamp(), ts.nanosecond());
-    let later = FileTime::from_unix_time(ts.unix_timestamp() as i64 + 3600, ts.nanosecond() as u32);
+    let later = FileTime::from_unix_time(ts.unix_timestamp() + 3600, ts.nanosecond());
     filetime::set_file_times(at.plus_as_string(file_a), now, now).unwrap();
     filetime::set_file_times(at.plus_as_string(file_b), now, later).unwrap();
